Refactor services to more logical structure
